+++ tests/utils/__init__.py
@@ -14,3 +14,16 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+import json
+from os import path
+
+FIXTURES_DIR = "tests/fixtures"
+
+
+def read_fixture(fixture_file_name):
+    with open(path.join(FIXTURES_DIR, fixture_file_name), "rb") as fixture_file:
+        return fixture_file.read()
+
+
+def load_fixture(fixture_file_name):
+    return json.loads(read_fixture(fixture_file_name))
